pyproject.toml
[project]
name = "monzo-lunch-money"
version = "0.1.0"
description = "Scripts to push monzo transactions to lunch money"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "loguru==0.7.2",
  "pandas==2.2.3",
  "sqlalchemy==2.0.36",
  "lunchable==1.3.3",
  "requests==2.32.3",
  "numpy>=2.1.0",
  "python-dotenv>=1.0.1",
  "psycopg2>=2.9.10",
  "ipykernel>=6.29.5",
  "ntfy-wrapper>=0.1.7",
  "monzo-api >= 1.1.0",
  "monzo_api_wrapper >= 0.1.1",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.2.0",
    "pre-commit",
    "mypy>=0.991",
    "ruff>=0.7.3",
    "pandas-stubs>=2.2.2.240807",
    "sqlalchemy-stubs>=0.4",
    "loguru-mypy>=0.0.4",
    "doq>=0.10.0",
    "types-requests>=2.32.0.20241016",
]

[tool.mypy]
files = ["monzo_lunch_money"]
disallow_untyped_defs = true
disallow_any_unimported = false
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["src/tests"]

[tool.ruff]
target-version = "py39"
line-length = 100
fix = true

[tool.ruff.lint]
pydocstyle.convention = "google"
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
    # docstrings
    "D",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
    # docstrings blank line spacing
    "D205",
    # Missing docstring in public module
    "D100",
    # Missing docstring in public package
    "D104",
]

[tool.ruff.lint.per-file-ignores]
"src/tests/*" = ["S101", "D103"]

[tool.ruff.format]
preview = true

[tool.doq]
formatter = "google"

[tool.docformatter]
recursive = true
wrap-summaries = 88
wrap-descriptions = 81
blank = true
